Don't build selenium with heavy machinery (cowbuilder + reprepro)
Squashed commit of the following:

commit 29601ecf4ae719cce80fa5606e237a863db59ded
Author: Olivier Berger <olivier.berger@telecom-sudparis.eu>
Date:   Wed Apr 30 09:25:52 2014 +0200

    Only rebuild selenium if it isn't installed already

commit 6f8d8c98c74263148544768dff5fea4af84dd1af
Author: Olivier Berger <olivier.berger@telecom-sudparis.eu>
Date:   Tue Apr 29 19:00:10 2014 +0200

    Get rid of cowbuilder and reprepro to just build a package in the
    3rd-party subdir

commit 6b3f32ef7d9161a5b036807f051184d6b1471162
Author: Olivier Berger <olivier.berger@telecom-sudparis.eu>
Date:   Tue Apr 29 18:56:04 2014 +0200

    Instead of an external Makefile, use the "classical" get-orig-source
    Debian way

+++ vm/scripts/run-testsuite.sh
@@ -1,19 +1,17 @@
 #!/bin/bash
 
-# Selenium dependencies
-aptitude -y install default-jre iceweasel
 
-# Build selenium
-aptitude -y install cowbuilder
-mkdir -p ~/builder/cow/
-DISTROLIST=wheezy /usr/src/fusionforge/tests/scripts/manage-cowbuilder.sh
+# Build an unofficial package for selenium and install it
+if ! dpkg -l selenium | grep -q ^ii ; then
+    cd /usr/src/fusionforge/3rd-party/selenium/selenium
+    debian/rules get-orig-source
+    debuild --no-lintian --no-tgz-check -us -uc
+    dpkg -i /usr/src/fusionforge/3rd-party/selenium/selenium_*_all.deb
 
-mkdir -p /usr/src/build/debian/conf/
-aptitude -y install reprepro
-echo -e "Codename: wheezy\nArchitectures: amd64 source\nComponents: main" > /usr/src/build/debian/conf/distributions
-cd /usr/src/fusionforge/3rd-party/selenium
-make
-dpkg -i /usr/src/build/debian/pool/main/s/selenium/selenium_*_all.deb
+    # Selenium dependencies
+    aptitude -y install default-jre iceweasel
+
+fi
 
 
 # Test dependencies
